feat(backend): check fingerprint mismatch and fix misspelling in InvalidRefreshToken exception name

+++ backend/src/auth/service.py
@@ -3,13 +3,13 @@ from datetime import UTC, datetime, timedelta
 
 from src.auth.exceptions import (
     CookieError,
-    InvalidRerfreshToken,
+    InvalidRefreshToken,
     RefreshSessionNotFound,
     TokenExpired,
     UserAuthDenied,
 )
 from src.auth.repository import AuthRepository
-from src.auth.schemas import RefreshSessionCreate, Tokens
+from src.auth.schemas import RefreshSessionCreate, RefreshSessionUpdate, Tokens
 from src.auth.utils import generate_jwt_token
 from src.core.exceptions import NotFound
 from src.users.models import User
@@ -75,18 +75,22 @@ class AuthService:
             access_token=access_token, refresh_token=refresh_session.refresh_token
         )
 
-    async def refresh_token(self, refresh_token: uuid.UUID) -> Tokens:
+    async def refresh_token(self, data: RefreshSessionUpdate) -> Tokens:
         try:
-            session = await self.repository.get_refresh_session(refresh_token)
+            session = await self.repository.get_refresh_session(data.refresh_token)
         except RefreshSessionNotFound:
-            raise InvalidRerfreshToken
+            raise InvalidRefreshToken
+
+        if session.fingerprint != data.fingerprint:
+            raise InvalidRefreshToken
 
         await self.repository.delete(session)
 
         if datetime.now(UTC).timestamp() >= session.expires_in:
-            raise TokenExpired
+            raise InvalidRefreshToken
 
         expires_in = int((datetime.now(UTC) + timedelta(weeks=4)).timestamp())
+
         new_session = await self.repository.create_refresh_session(
             RefreshSessionCreate(
                 user_id=session.user_id,
